+++ package.json
@@ -27,6 +27,7 @@
     "express": "^4.16.4",
     "express-basic-auth": "^1.2.0",
     "handlebars": "^4.1.2",
+    "ky": "^0.15.0",
     "lodash": "^4.17.11",
     "marked": "^0.7.0",
     "mkdirp": "^0.5.1",
